// backend/src/modules/progress/interfaces/progress-type.enum.ts
// src/modules/progress/interfaces/progress-type.enum.ts
export enum ProgressType {
  PRONUNCIATION = 'pronunciation',
  VOCABULARY = 'vocabulary',
  GRAMMAR = 'grammar',
  COMPREHENSION = 'comprehension',
  FLUENCY = 'fluency',
  ARTICULATION = 'articulation',
  PHONOLOGY = 'phonology',
  LANGUAGE = 'language',
  SOCIAL_COMMUNICATION = 'social_communication',
  OVERALL = 'overall',
}